backend/requirements.txt
PlexAPI
Flask
python-dotenv
cachetools